updated softmaxes
